Move metrics package to backend
