+++ internal/workflow/cache/cache.go
@@ -4,10 +4,10 @@ import (
 	"context"
 	"time"
 
+	"github.com/cschleiden/go-workflows/backend/metrics"
 	"github.com/cschleiden/go-workflows/core"
 	"github.com/cschleiden/go-workflows/internal/metrickeys"
 	"github.com/cschleiden/go-workflows/internal/workflow"
-	"github.com/cschleiden/go-workflows/metrics"
 	"github.com/jellydator/ttlcache/v3"
 )
